Refactor menus into separate modules
